Refactor tests + add tests for faker.fake
This commit splits the HelixSpec tests into individual files for each method.
This commit also adds tests against `Faker.js`'s `faker.fake()` method
for easy computed value generating!
